backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.116.1
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0

langchain==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
langsmith==0.4.8
langchain-openai==0.3.28
langchain-community==0.3.27

multidict==6.6.3
orjson==3.11.1
packaging==25.0
propcache==0.3.2

pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1

PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2

tavily-python==0.7.10

tenacity==9.1.2
tiktoken==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0

uv==0.8.3
uvicorn==0.35.0

uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.23.0

gradio==5.38.2
gradio-client==1.11.0